Move inject test config logic into Injector
